Move ExecutorStatus abstraction into the only module it is used in
