+++ parsl/dataflow/job_error_handler.py
@@ -1,6 +1,6 @@
 from typing import List, Dict
 
-from parsl.dataflow.executor_status import ExecutorStatus
+from parsl.dataflow.job_status_poller import ExecutorStatus
 from parsl.executors.base import ParslExecutor
 from parsl.providers.base import JobStatus, JobState
