+++ parsl/dataflow/strategy.py
@@ -4,7 +4,7 @@ import math
 import warnings
 from typing import Dict, List, Optional
 
-from parsl.dataflow.executor_status import ExecutorStatus
+from parsl.dataflow.job_status_poller import ExecutorStatus
 from parsl.executors import HighThroughputExecutor
 from parsl.executors.base import ParslExecutor
 from parsl.executors.status_handling import BlockProviderExecutor
